requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Vector and sparse retrieval
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
sentence-transformers>=2.2.0

# Graph processing
networkx>=3.1

# NLP and entity recognition
spacy>=3.5.0
# Run: python -m spacy download en_core_web_sm

# LLM and API clients
requests>=2.31.0
pyyaml>=6.0

# Ollama integration
ollama>=0.1.0

# Evaluation metrics
rouge-score>=0.1.2
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
